requirements.txt
aiohttp==3.1.3
async-timeout==2.0.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
haxor==1.2.1
idna==2.7
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
multidict==4.4.1
newspaper3k==0.2.7
nltk==3.3
Pillow==5.2.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
six==1.11.0
SQLAlchemy==1.2.11
tinysegmenter==0.3
tldextract==2.2.0
urllib3==1.23
Werkzeug==0.14.1
yarl==1.2.6
